+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main
 
+- Added new metrics instrumentation for Node 21. ([#1178](https://github.com/heroku/heroku-buildpack-nodejs/pull/1178))
+
 ## v231 (2023-12-01)
 
 - Updated metrics plugin to v15. ([#1184](https://github.com/heroku/heroku-buildpack-nodejs/pull/1184))